Ready to Push FB Login first try
Got the website to show your name after logging in..
Push to github, build on phonegap, and test :)

Sidenote, most of the night was spent looking up how to
create an android key hash. I ended up installing JDK
and used keytool to create a private key in a keystore
called appuki.keystore in Program Files/Java/jdk ... /
The alias is asifrc and the passwords are the oldschool one
with the first letter capitalized. I entered this into
Phonegap Build, and I just unlock it within an hour before
building. This will be the first test to see if it works.
Bismillah :)